extracted emails to a module
